+++ weli_writer_android/app/src/main/java/com/szoldapps/weli/writer/data/db/mapper/RoundMapper.kt
@@ -11,7 +11,7 @@ fun RoundEntity.mapToRound(): Round =
 
 fun List<RoundEntity>.mapToRounds(): List<Round> = this.map { gameEntity -> gameEntity.mapToRound() }
 
-fun Round.mapToRoundEntity(gameId: Int): RoundEntity =
+fun Round.mapToRoundEntity(gameId: Long): RoundEntity =
     RoundEntity(
         dateTime = date,
         gameId = gameId
