+++ weli_writer_android/app/src/main/java/com/szoldapps/weli/writer/domain/WeliRepository.kt
@@ -6,17 +6,17 @@ interface WeliRepository {
 
     val matches: LiveData<List<Match>>
 
-    fun gamesByMatchId(matchId: Int): LiveData<List<Game>>
+    fun gamesByMatchId(matchId: Long): LiveData<List<Game>>
 
-    fun roundsByGameId(gameId: Int): LiveData<List<Round>>
+    fun roundsByGameId(gameId: Long): LiveData<List<Round>>
 
     fun roundValuesByRoundId(roundId: Long): LiveData<List<RoundValue>>
 
     suspend fun addMatch(match: Match)
 
-    suspend fun addGame(game: Game, matchId: Int)
+    suspend fun addGame(game: Game, matchId: Long)
 
-    suspend fun addRound(round: Round, gameId: Int)
+    suspend fun addRound(round: Round, gameId: Long)
 
     suspend fun addRoundValue(roundValue: RoundValue, roundId: Long)
 }
